+++ CHANGELOG.md
@@ -40,6 +40,7 @@ sudo service mycodoflask restart
  - Add Input: ENS160 ([#1434](https://github.com/kizniche/Mycodo/pull/1434))
  - Add Output: Waveshare 8-Channel Raspberry Pi Relay Board B ([#1434](https://github.com/kizniche/Mycodo/pull/1434))
  - Add Function: Camera: rpicam ([#1487](https://github.com/kizniche/Mycodo/pull/1487))
+ - Add Input: VEML7700 Ambient Light Sensor ([#1491](https://github.com/kizniche/Mycodo/pull/1491))
 
 ### Miscellaneous